+++ backend/base.py
@@ -1,4 +1,5 @@
-from flask import Flask
+from flask import Flask, jsonify
+from database import load_jobs_from_db
 
 api = Flask(__name__)
 
@@ -10,6 +11,11 @@ def my_profile():
     }
 
     return response_body
+
+@api.route('/list_jobs')
+def list_jobs():
+    jobs = load_jobs_from_db()
+    return jsonify(jobs)
 
 if __name__ == '__main__':
     api.run(host='0.0.0.0', debug=True)
