+++ src/App.js
@@ -1,58 +1,19 @@
-import { useState } from 'react'
-import axios from "axios";
-import logo from './logo.svg';
-import './App.css';
-
+import Home from "./pages1/Home";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import CreatePost from "./pages1/CreatePost";
+import { routhPath } from "./routes1/routes";
+import AllPost from "./pages1/Allpost";
+import CV from "./pages1/cvrank.js"
 function App() {
-
-   // new line start
-  const [profileData, setProfileData] = useState(null)
-
-  function getData() {
-    axios({
-      method: "GET",
-      url:"/profile",
-    })
-    .then((response) => {
-      const res =response.data
-      setProfileData(({
-        profile_name: res.name,
-        about_me: res.about}))
-    }).catch((error) => {
-      if (error.response) {
-        console.log(error.response)
-        console.log(error.response.status)
-        console.log(error.response.headers)
-        }
-    })}
-    //end of new line 
-
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-
-        {/* new line start*/}
-        <p>To get your profile details: </p><button onClick={getData}>Click me</button>
-        {profileData && <div>
-              <p>Profile name: {profileData.profile_name}</p>
-              <p>About me: {profileData.about_me}</p>
-            </div>
-        }
-         {/* end of new line */}
-      </header>
-    </div>
+    <Router>
+      <Routes>
+        <Route path={routhPath.home} element={<Home />} />
+        <Route path={routhPath.create} element={<CreatePost />} />
+        <Route path={routhPath.post} element={<AllPost />} />
+        <Route path={routhPath.Cvrank} element={<CV />}/>
+      </Routes>
+    </Router>
   );
 }
